answer_evaluation_app/requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
fpdf2==2.7.6
reportlab==4.0.7
chromadb==0.4.18
sentence-transformers==2.7.0
huggingface-hub==0.24.0
scikit-learn==1.3.2
numpy==1.26.2
werkzeug==3.0.1

# DeepSeek-OCR dependencies (GPU CUDA 12.1 version)
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
transformers==4.45.0
accelerate==0.34.0
addict==2.4.0
timm==0.9.12
einops==0.7.0
